tf-idf the query
changed search to ranked retrieval instead of boolean

cosine similarity

jaccard coefficient is weighed in

index the index using metadata.json

+++ posting.py
@@ -1,7 +1,7 @@
 class Posting:
     __slots__ = ['docid', 'tfidf', 'url'] 
     
-    def __init__(self, docid: int, tfidf: int, url: str = "none"):
+    def __init__(self, docid: int, tfidf: float, url: str = "none"):
         self.docid = docid
         self.tfidf = tfidf
         self.url = url
